refactoring: put FAB/Modal into separate components (to make it easier to work with it)
