+++ frontend/src/app/features/modal/fab-integration/fab-integration.component.html
@@ -1,0 +1,2 @@
+<app-fab [buttons]="buttons"></app-fab>
+<app-rit-create-modal></app-rit-create-modal>
