+++ frontend/src/app/features/modal/rit-create-modal/rit-create-modal.component.spec.ts
@@ -1,0 +1,203 @@
+import { provideHttpClient } from '@angular/common/http';
+import { ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { IonicModule } from '@ionic/angular';
+import { ActionSheetController, ToastController } from '@ionic/angular/standalone';
+import { of, throwError } from 'rxjs';
+import { RitService } from '../../../shared/services/rit.service';
+import { UserService } from '../../../shared/services/user.service';
+import { RitCreateComponent } from '../../rit/rit-create/rit-create.component';
+import { RitCreateModalComponent } from './rit-create-modal.component';
+const userServiceMock = {
+  isLoggedIn: () => of(true)
+};
+
+describe('RitCreateModalComponent', () => {
+  let component: RitCreateModalComponent;
+  let fixture: ComponentFixture<RitCreateModalComponent>;
+  let ritServiceSpy: jasmine.SpyObj<RitService>;
+  let toastControllerSpy: jasmine.SpyObj<ToastController>;
+  let actionSheetControllerSpy: jasmine.SpyObj<ActionSheetController>;
+
+  beforeEach(async () => {
+    ritServiceSpy = jasmine.createSpyObj('RitService', ['createRit', 'getRits', 'getRitsErrorStream']);
+    ritServiceSpy.getRits.and.returnValue(of([]));
+    ritServiceSpy.getRitsErrorStream.and.returnValue(of({}));
+    const toastSpy = jasmine.createSpyObj('ToastController', ['create']);
+    const actionSheetSpy = jasmine.createSpyObj('ActionSheetController', ['create']);
+
+    TestBed.configureTestingModule({
+      imports: [RitCreateModalComponent,IonicModule.forRoot(),
+        NoopAnimationsModule],
+        providers: [
+          { provide: UserService, useValue: userServiceMock },
+          { provide: RitService, useValue: ritServiceSpy },
+          { provide: ToastController, useValue: toastSpy },
+          { provide: ActionSheetController, useValue: actionSheetSpy },
+          provideHttpClient()
+        ]
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(RitCreateModalComponent);
+    component = fixture.componentInstance;
+    ritServiceSpy = TestBed.inject(RitService) as jasmine.SpyObj<RitService>;
+    toastControllerSpy = TestBed.inject(ToastController) as jasmine.SpyObj<ToastController>;
+    actionSheetControllerSpy = TestBed.inject(ActionSheetController) as jasmine.SpyObj<ActionSheetController>;
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+
+  it('should call createRit and show success toast on success', fakeAsync(() => {
+    const toast = { present: jasmine.createSpy() };
+    toastControllerSpy.create.and.returnValue(Promise.resolve(toast as any));
+
+    component.modal = { dismiss: jasmine.createSpy() } as any;
+
+    component.ritCreateComponent = {
+      ritName: 'Test Rit',
+      details: 'Details',
+      tags: ['tag1']
+    } as any;
+
+    ritServiceSpy.createRit.and.returnValue(of({}));
+
+    component.confirm();
+    tick();
+
+    expect(ritServiceSpy.createRit).toHaveBeenCalled();
+    expect(component.modal.dismiss).toHaveBeenCalledWith(null, 'confirm');
+    expect(toastControllerSpy.create).toHaveBeenCalledWith(jasmine.objectContaining({ message: 'Rit created successfully!' }));
+    expect(toastControllerSpy.create).toHaveBeenCalledWith(jasmine.objectContaining({ color: 'success' }));
+    expect(toast.present).toHaveBeenCalled();
+  }));
+
+  it('should call createRit and show error toast on unknown error', fakeAsync(() => {
+    const toast = { present: jasmine.createSpy() };
+    toastControllerSpy.create.and.returnValue(Promise.resolve(toast as any));
+
+    component.modal = { dismiss: jasmine.createSpy() } as any;
+
+    component.ritCreateComponent = {
+      ritName: 'test rit',
+      details: 'Details',
+      tags: ['tag1'],
+      ritNameErrorMessage: '',
+      tagsErrorMessage: '',
+      detailsErrorMessage: ''
+    } as RitCreateComponent;
+
+    const mockErrorResponse = {
+      error: {
+        error: 'Unknown error',
+      }
+    };
+
+    ritServiceSpy.createRit.and.returnValue(throwError(() => mockErrorResponse));
+
+    component.confirm();
+    tick();
+
+    expect(toastControllerSpy.create).toHaveBeenCalledWith(jasmine.objectContaining({ color: 'danger', message: 'Unknown error' }));
+    expect(toast.present).toHaveBeenCalled();
+    expect(component.ritCreateComponent.ritNameErrorMessage).toEqual('');
+    expect(component.ritCreateComponent.tagsErrorMessage).toEqual('');
+    expect(component.ritCreateComponent.detailsErrorMessage).toEqual('');
+  }));
+
+  it('should call createRit and not show error toast on field errors', fakeAsync(() => {
+    const toast = { present: jasmine.createSpy() };
+    toastControllerSpy.create.and.returnValue(Promise.resolve(toast as any));
+
+    component.modal = { dismiss: jasmine.createSpy() } as any;
+
+    component.ritCreateComponent = {
+      ritName: undefined,
+      details: 'Details',
+      tags: ['tag1'],
+      ritNameErrorMessage: '',
+      tagsErrorMessage: '',
+      detailsErrorMessage: ''
+    } as RitCreateComponent;
+
+    const mockErrorResponse = {
+      error: {
+        error: 'Validation failed',
+        fields: {
+          name: [
+            'must not be empty'
+          ],
+          details: [
+            'test'
+          ],
+          tags: [
+            'test'
+          ]
+        }
+      }
+    };
+
+    ritServiceSpy.createRit.and.returnValue(throwError(() => mockErrorResponse));
+
+    component.confirm();
+    tick();
+
+    expect(toastControllerSpy.create).not.toHaveBeenCalledWith(jasmine.objectContaining({ color: 'danger', message: 'Unknown error' }));
+    expect(toast.present).not.toHaveBeenCalled();
+    expect(component.ritCreateComponent.ritNameErrorMessage).toEqual('must not be empty');
+    expect(component.ritCreateComponent.tagsErrorMessage).toEqual('test');
+    expect(component.ritCreateComponent.detailsErrorMessage).toEqual('test');
+  }));
+
+  it('should call modal.dismiss with null and "cancel" when Cancel button logic is used', async () => {
+    let dismissCalled = false;
+    let receivedData = null;
+    let receivedRole = '';
+
+    component.modal = {
+      dismiss: async (data: any, role: string) => {
+        dismissCalled = true;
+        receivedData = data;
+        receivedRole = role;
+      }
+    } as any;
+
+    await component.modal.dismiss(null, 'cancel');
+
+    expect(dismissCalled).toBeTrue();
+    expect(receivedData).toBeNull();
+    expect(receivedRole).toBe('cancel');
+  });
+
+  it('canDismiss returns true if role is not cancel', async () => {
+    const result = await component.canDismiss({}, 'confirm');
+    expect(result).toBeTrue();
+  });
+
+  it('canDismiss returns true if actionSheet returns confirm', async () => {
+    const present = jasmine.createSpy().and.returnValue(Promise.resolve());
+    const onWillDismiss = jasmine.createSpy().and.returnValue(Promise.resolve({ role: 'confirm' }));
+    actionSheetControllerSpy.create.and.returnValue(Promise.resolve({
+      present,
+      onWillDismiss
+    } as any));
+
+    const result = await component.canDismiss({}, 'cancel');
+    expect(result).toBeTrue();
+    expect(present).toHaveBeenCalled();
+    expect(onWillDismiss).toHaveBeenCalled();
+  });
+
+  it('canDismiss returns false if actionSheet returns cancel', async () => {
+    const present = jasmine.createSpy().and.returnValue(Promise.resolve());
+    const onWillDismiss = jasmine.createSpy().and.returnValue(Promise.resolve({ role: 'cancel' }));
+    actionSheetControllerSpy.create.and.returnValue(Promise.resolve({
+      present,
+      onWillDismiss
+    } as any));
+
+    const result = await component.canDismiss({}, 'cancel');
+    expect(result).toBeFalse();
+  });
+});
