+++ frontend/src/app/features/tabs/home/home.component.html
@@ -8,8 +8,7 @@
 
   <div class="rits-header-row">
     <ion-label>Your latest Rits</ion-label>
-    <ion-button size="small" (click)="goToRitsTab()" (keydown.enter)="goToRitsTab()" class="show-all-button"
-      data-testid="show-all-rits">
+    <ion-button size="small" (click)="goToRitsTab()" (keydown.enter)="goToRitsTab()" class="show-all-button" data-testid="show-all-rits">
       Show all
       <ion-icon slot="end" name="chevron-forward"></ion-icon>
     </ion-button>
@@ -23,41 +22,6 @@
     <app-rit-list-item *ngFor="let rit of latestRits()" [rit]="rit">
     </app-rit-list-item>
   </ion-list>
-
-  <ion-fab data-testid="rit-create-button" slot="fixed" vertical="bottom" horizontal="end">
-    <ion-fab-button>
-      <ion-icon name="add-outline"></ion-icon>
-    </ion-fab-button>
-    <ion-fab-list side="top">
-      <ion-fab-button id="open-modal">
-        <ion-icon name="add-outline"></ion-icon>
-      </ion-fab-button>
-    </ion-fab-list>
-  </ion-fab>
-
-  <ion-modal #modal trigger="open-modal" [canDismiss]="canDismiss" [initialBreakpoint]="0.8" [breakpoints]="[0, 0.8]">
-    <ng-template>
-      <ion-header>
-        <ion-toolbar>
-          <ion-buttons slot="start">
-            <ion-button (click)="modal.dismiss(null, 'cancel')" (keydown.enter)="modal.dismiss(null, 'cancel')">
-              Cancel
-            </ion-button>
-          </ion-buttons>
-          <ion-title>Rit Create</ion-title>
-          <ion-buttons slot="end">
-            <ion-button [disabled]="!ritCreate.ritName?.trim()" (click)="confirm()" (keydown.enter)="confirm()"
-              [strong]="true">
-              Confirm
-            </ion-button>
-          </ion-buttons>
-        </ion-toolbar>
-      </ion-header>
-
-      <ion-content class="ion-padding">
-        <app-rit-create #ritCreate></app-rit-create>
-      </ion-content>
-    </ng-template>
-  </ion-modal>
-
 </ion-content>
+
+<app-fab-integration></app-fab-integration>
