+++ frontend/src/app/features/tabs/home/home.component.spec.ts
@@ -61,180 +61,6 @@ describe('HomeComponent', () => {
     expect(title?.textContent).toContain('Home');
   });
 
-  it('should call createRit and show success toast on success', fakeAsync(() => {
-    const toast = { present: jasmine.createSpy() };
-    toastControllerSpy.create.and.returnValue(Promise.resolve(toast as any));
-
-    component.modal = { dismiss: jasmine.createSpy() } as any;
-
-    component.ritCreateComponent = {
-      ritName: 'Test Rit',
-      details: 'Details',
-      tags: ['tag1']
-    } as any;
-
-    ritServiceSpy.createRit.and.returnValue(of({}));
-
-    component.confirm();
-    tick();
-
-    expect(ritServiceSpy.createRit).toHaveBeenCalled();
-    expect(component.modal.dismiss).toHaveBeenCalledWith(null, 'confirm');
-    expect(toastControllerSpy.create).toHaveBeenCalledWith(jasmine.objectContaining({ message: 'Rit created successfully!' }));
-    expect(toastControllerSpy.create).toHaveBeenCalledWith(jasmine.objectContaining({ color: 'success' }));
-    expect(toast.present).toHaveBeenCalled();
-  }));
-
-  it('should call createRit and show error toast on unknown error', fakeAsync(() => {
-    const toast = { present: jasmine.createSpy() };
-    toastControllerSpy.create.and.returnValue(Promise.resolve(toast as any));
-
-    component.modal = { dismiss: jasmine.createSpy() } as any;
-
-    component.ritCreateComponent = {
-      ritName: 'test rit',
-      details: 'Details',
-      tags: ['tag1'],
-      ritNameErrorMessage: '',
-      tagsErrorMessage: '',
-      detailsErrorMessage: ''
-    } as RitCreateComponent;
-
-    const mockErrorResponse = {
-      error: {
-        error: 'Unknown error',
-      }
-    };
-
-    ritServiceSpy.createRit.and.returnValue(throwError(() => mockErrorResponse));
-
-    component.confirm();
-    tick();
-
-    expect(toastControllerSpy.create).toHaveBeenCalledWith(jasmine.objectContaining({ color: 'danger', message: 'Unknown error' }));
-    expect(toast.present).toHaveBeenCalled();
-    expect(component.ritCreateComponent.ritNameErrorMessage).toEqual('');
-    expect(component.ritCreateComponent.tagsErrorMessage).toEqual('');
-    expect(component.ritCreateComponent.detailsErrorMessage).toEqual('');
-  }));
-
-  it('should call createRit and not show error toast on field errors', fakeAsync(() => {
-    const toast = { present: jasmine.createSpy() };
-    toastControllerSpy.create.and.returnValue(Promise.resolve(toast as any));
-
-    component.modal = { dismiss: jasmine.createSpy() } as any;
-
-    component.ritCreateComponent = {
-      ritName: undefined,
-      details: 'Details',
-      tags: ['tag1'],
-      ritNameErrorMessage: '',
-      tagsErrorMessage: '',
-      detailsErrorMessage: ''
-    } as RitCreateComponent;
-
-    const mockErrorResponse = {
-      error: {
-        error: 'Validation failed',
-        fields: {
-          name: [
-            'must not be empty'
-          ],
-          details: [
-            'test'
-          ],
-          tags: [
-            'test'
-          ]
-        }
-      }
-    };
-
-    ritServiceSpy.createRit.and.returnValue(throwError(() => mockErrorResponse));
-
-    component.confirm();
-    tick();
-
-    expect(toastControllerSpy.create).not.toHaveBeenCalledWith(jasmine.objectContaining({ color: 'danger', message: 'Unknown error' }));
-    expect(toast.present).not.toHaveBeenCalled();
-    expect(component.ritCreateComponent.ritNameErrorMessage).toEqual('must not be empty');
-    expect(component.ritCreateComponent.tagsErrorMessage).toEqual('test');
-    expect(component.ritCreateComponent.detailsErrorMessage).toEqual('test');
-  }));
-
-  it('should call modal.dismiss with null and "cancel" when Cancel button logic is used', async () => {
-    let dismissCalled = false;
-    let receivedData = null;
-    let receivedRole = '';
-
-    component.modal = {
-      dismiss: async (data: any, role: string) => {
-        dismissCalled = true;
-        receivedData = data;
-        receivedRole = role;
-      }
-    } as any;
-
-    await component.modal.dismiss(null, 'cancel');
-
-    expect(dismissCalled).toBeTrue();
-    expect(receivedData).toBeNull();
-    expect(receivedRole).toBe('cancel');
-  });
-
-  it('canDismiss returns true if role is not cancel', async () => {
-    const result = await component.canDismiss({}, 'confirm');
-    expect(result).toBeTrue();
-  });
-
-  it('canDismiss returns true if actionSheet returns confirm', async () => {
-    const present = jasmine.createSpy().and.returnValue(Promise.resolve());
-    const onWillDismiss = jasmine.createSpy().and.returnValue(Promise.resolve({ role: 'confirm' }));
-    actionSheetControllerSpy.create.and.returnValue(Promise.resolve({
-      present,
-      onWillDismiss
-    } as any));
-
-    const result = await component.canDismiss({}, 'cancel');
-    expect(result).toBeTrue();
-    expect(present).toHaveBeenCalled();
-    expect(onWillDismiss).toHaveBeenCalled();
-  });
-
-  it('canDismiss returns false if actionSheet returns cancel', async () => {
-    const present = jasmine.createSpy().and.returnValue(Promise.resolve());
-    const onWillDismiss = jasmine.createSpy().and.returnValue(Promise.resolve({ role: 'cancel' }));
-    actionSheetControllerSpy.create.and.returnValue(Promise.resolve({
-      present,
-      onWillDismiss
-    } as any));
-
-    const result = await component.canDismiss({}, 'cancel');
-    expect(result).toBeFalse();
-  });
-
-  it('should render ion-fab-button when logged in', async () => {
-    fixture.componentInstance.ionViewWillEnter();
-    fixture.detectChanges();
-
-    const fabButton = fixture.nativeElement.querySelector('[data-testid="rit-create-button"]');
-    expect(fabButton).toBeTruthy();
-  });
-
-  it('should not render ion-fab-button when not logged in', async () => {
-    userServiceMock.isLoggedIn = () => of(false);
-
-    fixture = TestBed.createComponent(HomeComponent);
-    component = fixture.componentInstance;
-    fixture.componentInstance.ionViewWillEnter();
-    fixture.detectChanges();
-
-    const fabButton = fixture.nativeElement.querySelector('[data-testid="rit-create-button"]');
-    expect(fabButton).toBeFalsy();
-
-    userServiceMock.isLoggedIn = () => of(true)
-  });
-
   it('should return the latest 10 rits sorted by lastInteractionAt descending', () => {
     const now = new Date().getTime();
     const ritsMock = Array.from({ length: 15 }, (_, i) => ({
